requirements.txt
networkx>=3.1
pyyaml>=6.0
requests>=2.31.0
streamlit>=1.24.0
pytest>=7.4.0
python-dotenv>=1.0.0
pyvis>=0.3.2
